+++ processor/src/main/java/arez/processor/ActionDescriptor.java
@@ -27,6 +27,7 @@ final class ActionDescriptor
   @Nonnull
   private final String _name;
   private final boolean _requireNewTransaction;
+  private final boolean _requireEnvironment;
   private final boolean _mutation;
   private final boolean _verifyRequired;
   private final boolean _reportParameters;
@@ -38,6 +39,7 @@ final class ActionDescriptor
   ActionDescriptor( @Nonnull final ComponentDescriptor componentDescriptor,
                     @Nonnull final String name,
                     final boolean requireNewTransaction,
+                    final boolean requireEnvironment,
                     final boolean mutation,
                     final boolean verifyRequired,
                     final boolean reportParameters,
@@ -47,6 +49,7 @@ final class ActionDescriptor
     _componentDescriptor = Objects.requireNonNull( componentDescriptor );
     _name = Objects.requireNonNull( name );
     _requireNewTransaction = requireNewTransaction;
+    _requireEnvironment = requireEnvironment;
     _mutation = mutation;
     _verifyRequired = verifyRequired;
     _reportParameters = reportParameters;
@@ -213,6 +216,14 @@ final class ActionDescriptor
     else
     {
       flags.add( "READ_ONLY" );
+    }
+    if ( _requireEnvironment )
+    {
+      flags.add( "ENVIRONMENT_REQUIRED" );
+    }
+    else
+    {
+      flags.add( "ENVIRONMENT_NOT_REQUIRED" );
     }
     if ( _verifyRequired )
     {