+++ processor/src/test/resources/expected/com/example/inject/Arez_MultipleArgsModel.java
@@ -116,7 +116,7 @@ public final class Arez_MultipleArgsModel extends MultipleArgsModel implements D
       Guards.apiInvariant( () -> ComponentState.isActive( this.$$arezi$$_state ), () -> "Method named 'myActionStuff' invoked on " + ComponentState.describe( this.$$arezi$$_state ) + " component named '" + $$arezi$$_name() + "'" );
     }
     try {
-      $$arezi$$_context().safeAction(Arez.areNamesEnabled() ? $$arezi$$_name() + ".myActionStuff" : null, () -> super.myActionStuff(), Flags.READ_WRITE | Flags.VERIFY_ACTION_REQUIRED, null );
+      $$arezi$$_context().safeAction(Arez.areNamesEnabled() ? $$arezi$$_name() + ".myActionStuff" : null, () -> super.myActionStuff(), Flags.READ_WRITE | Flags.ENVIRONMENT_NOT_REQUIRED | Flags.VERIFY_ACTION_REQUIRED, null );
     } catch( final RuntimeException | Error $$arez_exception$$ ) {
       throw $$arez_exception$$;
     } catch( final Throwable $$arez_exception$$ ) {